requirements.txt
asgiref==3.8.1
Django==4.2.7
django-cors==0.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
pillow==11.1.0
psycopg2==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
sqlparse==0.5.3
python-decouple==3.8
